+++ src/mock/render.js
@@ -1,7 +1,7 @@
-import { RenderPosition } from '../view/render-data';
+import { RenderPosition } from './generate';
 import AbctractView from '../view/abstract-view';
 
-export const renderElement = (container, element, place) => {
+export function renderElement (container, element, place) {
   const parent = container instanceof AbctractView ? container.element : container;
   const child = element instanceof AbctractView ? element.element : element;
