+++ src/presenter/movie-list-presenter.js
@@ -8,51 +8,48 @@ import ExtraView from '../view/extra-view.js';
 import FooterView from '../view/footer-view.js';
 import EmtyListView from '../view/empty-list-view.js';
 import { renderElement } from '../mock/render.js';
-import { RenderPosition } from '../view/render-data.js';
-import { mainElement } from '../view/render-data.js';
+import { RenderPosition } from '../mock/generate.js';
+import MoviePresenter from './movie-presenter.js';
 
 const NEXT_POSTS_COUNT = 5;
 const footer = document.querySelector('.footer');
 
 export default class MovieListPresenter {
-#movieListContainer = null;
+#mainContainer = null;
 
 #cardsContainerComponent = new CardsContainerView();
 #cardsComponent = new CardsView();
-#menuComponent = new MenuView();
+menuComponent = new MenuView();
 #filterComponent = new FilterView();
 #buttonComponent = new ButtonView();
 #avatarComponent = new AvatarView();
 #extraComponent = new ExtraView();
 #footerComponent = new FooterView();
 #emptyListComponent = new EmtyListView();
+#cardsContainer = this.#cardsContainerComponent.element.querySelector('.films-list__container');
 
 #movies = [];
 
-constructor(movieListContainer) {
-  this.#movieListContainer = movieListContainer;
+constructor(mainContainer) {
+  this.#mainContainer = mainContainer;
 }
 
 init = (movies) => {
-  renderElement(mainElement, this.#menuComponent, RenderPosition.BEFOREEND);
-  this.#menuComponent.setActiveFilter(this.#emptyListComponent.element);
-  this.#menuComponent.setEmptyMessage(this.#emptyListComponent.element);
-
   this.#movies = [...movies];
+  renderElement(this.#mainContainer, this.menuComponent, RenderPosition.BEFOREEND);
+  this.menuComponent.setFiltersCount();
+  this.menuComponent.setActiveFilter(this.#emptyListComponent.element);
   this.#renderMovieList();
+
 }
 
 #renderSort = () => {
-  renderElement(mainElement, this.#filterComponent, RenderPosition.BEFOREEND);
+  renderElement(this.#mainContainer, this.#filterComponent, RenderPosition.BEFOREEND);
 }
 
 #renderMovie = (movie) => {
-  const cardsContainer = this.#cardsContainerComponent.element.querySelector('.films-list__container');
-  const cardComponent = new CardsView(movie);
-  // cardComponent.addClickHandler(movie, () => {
-  //   new PopupView().postClickHandler(movie, cardComponent);
-  // });
-  renderElement(cardsContainer, cardComponent, RenderPosition.BEFOREEND);
+  const moviePresenter = new MoviePresenter(this.#cardsContainer);
+  moviePresenter.init(movie);
 }
 
 #renderMovies = (from, to) => {
@@ -61,19 +58,19 @@ init = (movies) => {
 }
 
 #renderNoMovies = () => {
-  renderElement(this.#cardsContainerComponent, this.#filterComponent, RenderPosition.AFTERBEGIN);
-  // Метод для рендеринга заглушки
+  renderElement(this.#cardsContainer, this.#emptyListComponent, RenderPosition.BEFOREEND);
+  this.menuComponent.setEmptyMessage(this.#emptyListComponent.element);
 }
 
 #renderLoadMoreButton = () => {
-  renderElement(mainElement, this.#buttonComponent, RenderPosition.BEFOREEND);
+  renderElement(this.#mainContainer, this.#buttonComponent, RenderPosition.BEFOREEND);
   this.#buttonComponent.addClickControlsEvent();
 }
 
 #renderMovieList = () => {
   this.#renderSort();
 
-  renderElement(mainElement, this.#cardsContainerComponent, RenderPosition.BEFOREEND);
+  renderElement(this.#mainContainer, this.#cardsContainerComponent, RenderPosition.BEFOREEND);
 
   this.#renderMovies(0, Math.min(this.#movies.length, NEXT_POSTS_COUNT));
